requirements.txt
python-telegram-bot==20.7
instaloader==4.11
python-dotenv==1.0.0